dump: Fix Hostboot and Hardware dump offload
Recently, we consolidated our system entry path for all OpenPower
dumps. When the PHYP (Power Hypervisor) goes down, powervm-handler
restarts and we use the GetSubtreePaths method of the mapper for each
entry interface. Since dump object paths are linked to specific
interfaces—such as SBE, HostBoot, and Hardware—no entries are returned
when calling GetSubtreePaths on the System interface alone.

Therefore, we need to invoke mapper on SBE, HostBoot, and Hardware
interfaces.

Change-Id: I10968429e5e2fdc4634d7fde09f8708b84e5013d
Signed-off-by: SwethaParasa <parasa.swetha1@ibm.com>